Ignore new refs that mean deletion
On Bitbucket Server 3.2.2 when trying to delete a branch via a
plain `git push origin :branchname` we hit this plugin trying
to `CommitService.getCommit` on that deletion commit, which of
course fails with a `NoSuchCommitException` there.

When a from hash consists of 40 `0` a ref is being created,
but when a to hash consists of 40 `0` a ref is being deleted.

Fixes #14.